replace logout button with theme switch button
- create `ThemeSwitch` component for the theme switch button
- extract logout button to own component
- place logout component in user settings route
- add color transition to all elements
